separate the bar out as diff component add image make the turninput a textinput on browser
