+++ code/simulator/extensions/visualiser/Components/snapshot_parser.tsx
